Include multiple slope and ag land classes in solar_project.csv
This also prepares load_zone_ag_land_area.csv showing the total
amount of land in each class in each load zone, which can be used
to apply limits based on percentage of total land of each class
that can be used.

This uses a new all-inclusive script to do land use filtering for
utility-scale solar projects. The script does the same work as the
previous manual process: create a .csv file showing the amount of
solar-suitable area in each combination of solar cluster and nsrdb
grid cell. But it is now done via a Python script (using GRASS GIS),
and it now subdivides the results by Land Study Bureau agricultural
land class (A (none allowed), B, C or other) and slope class
(00_to_15 and 15_to_30 (percent)). This replaces the previous
process which only allowed slopes of 0 to 10 and erroneously used
degrees instead of percent (so it was roughly 0 to 17 degrees).

Work still needs to be done to import the new tables and attributes
into the data warehouse and export them for use in Switch models.